Update intro.mdx
I just wanted to tweak the grammar here and there to make it more understandable? In terms of translating to another language and for everyone (native and non native speakers) that understands English to understand it without having to think about it much. This way, the message seems a bit friendlier while still having a professional yet formal appeal to the message.