Update this page and make it precise and better
Whoever wrote this description for SKlauncher did a poor job. I've attempted to improve it, but I'm not familiar with this unfamiliar tool I'm writing in.

Here are the issues:
1- The text lacks clarity and precision.
2- The use of emojis in the features headings doesn't seem 
    appropriate (personal opinion).
3- It's incomplete. Why is "* - This feature is currently in 
    development" written at the end?
4- There are grammatical errors, such as punctuation.

If this was a test for a guide writer position at SKlauncher, please consider me. I have experience writing gaming guides and I'm a fast learner who consistently improves. However, if I have to work with this "black sheet" or whatever it is, please don't consider me. I prefer writing in a normal format without dealing with unfamiliar tools.